Flesh out attendances controller, create related views, start a preliminary competition views.
In this commit, the attendances controller now gets more completed
features. The views are still very basic and include just the
fundamentals. Tests have been added for the attendances, and a basic
leaderboard has been set up. Also, the main site layout is slightly
updated now, with a placeholder contact page and a beefier footer
partial.