Add groups/events page for user, and put more stats
Each user now has a page that shows the groups and events that they are in. In addition, some of the user's stats are on the views (ex. number of hours volunteered). Some statistics have been added to the group page. Also, the attendence-edit page now has an easy-unchecking button when appropriate.