Major styling/view changes
Many changes related to the views come in this commit. Bootstrap with
custom CSS is now very prevalent throughout most of the site.
Pagination is included for users and groups, and tables are now a part
of the site. Forms are better styled.

There are some changes to models, though; the group model has been
fixed to correctly destroy events upon deletion, events destroy
attendances, and deleting attendances will uncredit hours.