Add hours to memberships
Each membership now carries an hours attribute that represents the
number of hours a user has volunteered for a specific group. The views
now show this information, as well.

There is something to note, however: A particular test currently fails
due to a strange error involving the date attribute appearing as a
‘Fixnum’ in the tests  (but works fine in the actual view). Although
not ideal, this test is commented out and will be attended to later.