Add more tests for the display and removing group members
In this commit, a test has been added for checking things during removing members from a group. In addition, there are more tests for checking the info displayed on a group in multiple contexts. The index group test has been moved for organization purposes. Also, there are now 100 tests in total.