Include a test for uncrediting attendance on destruction
Test coverage reports show that the uncredit_hours feature of
the attendance model had not been tested.

In the process of making this test, something has been discovered:
The use of "destroy" instead of "destroy!" in the User model's leave
method has been causing complications in testing hours. Because of the
use of the former method, models have be reloaded before checking
their hour counts (as is the case in the new test). In fact, it may
explain some of the complications that have come up when testing for
hours in the other tests. The User methods will be changed later.