Quick fix for the late-night time issue
Some tests would fail after 11PM because the date/time values that
they would pass into models would be such that the test would be
trying to create an event that would span more than one day. However,
the models as they are specified now do not support multi-day events
and process the request as asking for an end time that's earlier than
the start time!

The models will have to be adjusted to support multi-day events in a
future release.